client: pull: Avoid panic by waiting for progress updates goroutine to finish
You can get a panic if the iteration through the progress updates finishes after PullModel returns and the progressWriter's underlying Writer is closed. You can simulate this by adding a time.Sleep(time.Second) between the updates and run a docker model pull - the pull will finish but the goroutine that handles progress updates will still be running and it will attempt to write to the progressWriter's underlying closed Writer.

Signed-off-by: Dorin Geman <dorin.geman@docker.com>